Moved statement related handler into dedicated package
