+++ server/src/main/java/com/consol/citrus/db/server/JdbcServer.java
@@ -22,21 +22,21 @@ import com.consol.citrus.db.server.controller.RuleBasedController;
 import com.consol.citrus.db.server.controller.SimpleJdbcController;
 import com.consol.citrus.db.server.exceptionhandler.JdbcServerExceptionHandler;
 import com.consol.citrus.db.server.handler.CloseCallableStatementHandler;
-import com.consol.citrus.db.server.handler.CloseStatementHandler;
 import com.consol.citrus.db.server.handler.CreateCallableStatementHandler;
-import com.consol.citrus.db.server.handler.CreatePreparedStatementHandler;
-import com.consol.citrus.db.server.handler.CreateStatementHandler;
 import com.consol.citrus.db.server.handler.ExecuteCallableStatementHandler;
-import com.consol.citrus.db.server.handler.ExecuteJsonQueryHandler;
-import com.consol.citrus.db.server.handler.ExecuteStatementHandler;
-import com.consol.citrus.db.server.handler.ExecuteUpdateHandler;
-import com.consol.citrus.db.server.handler.ExecuteXmlQueryHandler;
 import com.consol.citrus.db.server.handler.connection.CloseConnectionHandler;
 import com.consol.citrus.db.server.handler.connection.CommitTransactionStatementsHandler;
 import com.consol.citrus.db.server.handler.connection.GetTransactionStateHandler;
 import com.consol.citrus.db.server.handler.connection.OpenConnectionHandler;
 import com.consol.citrus.db.server.handler.connection.RollbackTransactionStatementsHandler;
 import com.consol.citrus.db.server.handler.connection.SetTransactionStateHandler;
+import com.consol.citrus.db.server.handler.statement.CloseStatementHandler;
+import com.consol.citrus.db.server.handler.statement.CreatePreparedStatementHandler;
+import com.consol.citrus.db.server.handler.statement.CreateStatementHandler;
+import com.consol.citrus.db.server.handler.statement.ExecuteJsonQueryHandler;
+import com.consol.citrus.db.server.handler.statement.ExecuteStatementHandler;
+import com.consol.citrus.db.server.handler.statement.ExecuteUpdateHandler;
+import com.consol.citrus.db.server.handler.statement.ExecuteXmlQueryHandler;
 import com.consol.citrus.db.server.transformer.JsonResponseTransformer;
 import com.consol.citrus.db.server.transformer.XmlResponseTransformer;
 import org.slf4j.Logger;
