+++ server/src/main/java/com/consol/citrus/db/server/handler/statement/ExecuteUpdateHandler.java
@@ -14,22 +14,22 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.db.server.handler;
+package com.consol.citrus.db.server.handler.statement;
 
 import com.consol.citrus.db.server.controller.JdbcController;
+import com.consol.citrus.db.server.handler.AbstractJdbcRequestHandler;
 import spark.Request;
 import spark.Response;
 
-public class ExecuteJsonQueryHandler extends AbstractJdbcRequestHandler {
+public class ExecuteUpdateHandler extends AbstractJdbcRequestHandler {
 
 
-    public ExecuteJsonQueryHandler(final JdbcController controller) {
+    public ExecuteUpdateHandler(final JdbcController controller) {
         super(controller);
     }
 
     @Override
     public Object handle(final Request request, final Response response) {
-        response.type("application/json");
-        return controller.executeQuery(request.body());
+        return controller.executeUpdate(request.body());
     }
 }
