+++ server/src/test/java/com/consol/citrus/db/server/handler/CloseStatementHandlerTest.java
@@ -17,6 +17,7 @@
 package com.consol.citrus.db.server.handler;
 
 import com.consol.citrus.db.server.controller.JdbcController;
+import com.consol.citrus.db.server.handler.statement.CloseStatementHandler;
 import org.testng.annotations.Test;
 import spark.Request;
 import spark.Response;
